# src/main/docs/api.yaml
openapi: 3.0.3
info:
  title: EventSync API
  description: API for event management and real-time participant engagement
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Development server
paths:
...
  /sessions:
    get:
      summary: List all sessions
      tags:
        - Sessions
      parameters:
        - name: room
          in: query
          schema:
            type: string
          description: Filter by room name
        - name: speaker
          in: query
          schema:
            type: string
          description: Filter by speaker name (matches first or last name)
        - name: live
          in: query
          schema:
            type: boolean
          description: Filter sessions currently live
        - name: event
          in: query
          schema:
            type: string
          description: Filter by event title
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number (starts at 1)
        - name: size
          in: query
          schema:
            type: integer
            default: 20
          description: Number of items per page
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionList'
    post:
      summary: Create a new session
      tags:
        - Sessions
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionInput'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          description: Unauthorized
        '404':
          description: Room or event not found
        '409':
          description: Session with this title already exists
...
  /sessions/{id}/questions:
    get:
      summary: List questions of a session
      tags:
        - Questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the session
        - name: sort
          in: query
          schema:
            type: string
            enum: [upvotes, createdAt]
            default: upvotes
          description: Sort order (upvotes descending, or createdAt ascending)
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number (starts at 1)
        - name: size
          in: query
          schema:
            type: integer
            default: 20
          description: Number of items per page
      responses:
        '200':
          description: List of questions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionList'
    post:
      summary: Ask a question on a session
      tags:
        - Questions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionInput'
      responses:
        '201':
          description: Question created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '400':
          description: Session not live - questions disabled
        '401':
          description: Unauthorized
  /sessions/{id}/questions/{qid}/upvote:
    post:
      summary: Upvote a question
      tags:
        - Questions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the session
        - name: qid
          in: path
          required: true
          schema:
            type: string
          description: ID of the question
      responses:
        '200':
          description: Question upvotée
          content:
            application/json:
              schema:
                type: integer
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EventRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        location:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        room:
          $ref: '#/components/schemas/Room'
        capacity:
          type: integer
        event:
          $ref: '#/components/schemas/EventRef'
        speakers:
          type: array
          items:
            $ref: '#/components/schemas/SpeakerRef'
        isLive:
          type: boolean
          description: True if the session is live
...
    SessionList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Session'
        meta:
          type: object
          properties:
            total:
              type: integer
            page:
              type: integer
            size:
              type: integer
    SessionInput:
      type: object
      required:
        - title
        - description
        - startDate
        - endDate
        - roomName
        - eventTitle
      properties:
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        roomName:
          type: string
        capacity:
          type: integer
        eventTitle:
          type: string
    Room:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
...
    SpeakerRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        profilePicture:
          type: string
          format: uri
        bio:
          type: string
...
    ParticipantRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
...
    Question:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        participant:
          $ref: '#/components/schemas/ParticipantRef'
        upvotes:
          type: integer
        isAnonymous:
          type: boolean
        createdAt:
          type: string
          format: date-time
    QuestionList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Question'
        meta:
          type: object
          properties:
            total:
              type: integer
            page:
              type: integer
            size:
              type: integer
...
    QuestionInput:
      type: object
      required:
        - content
      properties:
        title:
          type: string
        content:
          type: string
        participant:
          $ref: '#/components/schemas/ParticipantRef'
        isAnonymous:
          type: boolean
          default: false